Change log; Fix bugs
Former-commit-id: 42a785fcc9c8888853030246c1911582862de002 [formerly f5f0f5950d66c0985d762b9340332d4898144df0] [formerly b1e3dfbce90badc8b195bdf88fd6b8d5244eebf4 [formerly 42526ff46cb56210c4c9e62e17be91209cfde022]]
Former-commit-id: 3e7bbb016fad4c73223d8ef97cf84f81d74b1cd4 [formerly 1d8c914435a79f1124d71f3c2b8ffb669cfc2b17]
Former-commit-id: 02ae0bfabb6eae6acf1820a57a2c936eae94196f

+++ Frame/SDK/Core/Base/AFTime.hpp
@@ -232,6 +232,7 @@ public:
     {
         return mnTime;//GetFixTime();
     }
+
     int64_t Get(TimeType nType)
     {
         int year = 1970;
@@ -738,7 +739,7 @@ private:
             }
         }
 
-        return ret_.size();
+        return (int)ret_.size();
     }
 
 private:
